Basic implementation of installAndActivatePackage binder call
To make it easier to review this CL just implements the basic flow.
Following will be implemented in the follow-up CL(s):

* Error handling and clean up
* e2e tests

Since some binaries that are using an APEX that is being updated can run
in a separate mount namespace, we can't just destroy dm-verity and loop
devices that previously active version of APEX is mounted on top of.
Instead, we are making use of the DM_DEFERRED_REMOVE and
LO_FLAGS_AUTOCLEAR flags which will automatically tear down dm and loop
block devices respectively after last close.

Due to the same reasons, supporting a samegrade becomes a little bit
tricky: we need to pick up a name for the device that doesn't collide
with one that is backing a previous APEX that is still accessed by a
process in it's own mount namespace (same applies to the final location
of APEX on /data/apex/active).

To handle this a concept of "minor version" is introduced: name of a
dm-verity device now follows ${module_name}@{version_code}[_${minor}]
pattern. A minor version of the new dm-verity device is calculated as
max(minor versions of other dm-verity devices with same
${module_name}) + 1.

Test: atest ApexTestCases
Test: adb install --force-non-staged rebootless_apex_v2.apex
Bug: 187864524
Bug: 188713178
Merged-In: I346ecf1360be87320c718922130261f6f244f1ac
(cherry picked from commit 8cef66e2b9f3c2cbd6b8654fa8d12c16ee8687aa)

+++ apexd/apexd.cpp
@@ -15,6 +15,7 @@
  */
 
 #include "apexd.h"
+#include "apex_file_repository.h"
 #include "apexd_private.h"
 
 #include "apex_constants.h"
@@ -31,7 +32,6 @@
 #include "apexd_utils.h"
 #include "apexd_verity.h"
 #include "com_android_apex.h"
-#include "string_log.h"
 
 #include <ApexProperties.sysprop.h>
 #include <android-base/file.h>
@@ -78,10 +78,12 @@
 #include <queue>
 #include <sstream>
 #include <string>
+#include <string_view>
 #include <thread>
 #include <unordered_map>
 #include <unordered_set>
 
+using android::base::ConsumePrefix;
 using android::base::ErrnoError;
 using android::base::Error;
 using android::base::GetProperty;
@@ -213,8 +215,15 @@ std::unique_ptr<DmTable> CreateVerityTable(const ApexVerityData& verity_data,
 
 // Deletes a dm-verity device with a given name and path
 // Synchronizes on the device actually being deleted from userspace.
-Result<void> DeleteVerityDevice(const std::string& name) {
+Result<void> DeleteVerityDevice(const std::string& name, bool deferred) {
   DeviceMapper& dm = DeviceMapper::Instance();
+  if (deferred) {
+    if (!dm.DeleteDeviceDeferred(name)) {
+      return ErrnoError() << "Failed to issue deferred delete of verity device "
+                          << name;
+    }
+    return {};
+  }
   auto timeout = std::chrono::milliseconds(
       android::sysprop::ApexProperties::dm_delete_timeout().value_or(750));
   if (!dm.DeleteDevice(name, timeout)) {
@@ -250,7 +259,7 @@ class DmVerityDevice {
 
   ~DmVerityDevice() {
     if (!cleared_) {
-      Result<void> ret = DeleteVerityDevice(name_);
+      Result<void> ret = DeleteVerityDevice(name_, /* deferred= */ false);
       if (!ret.ok()) {
         LOG(ERROR) << ret.error();
       }
@@ -276,7 +285,7 @@ Result<DmVerityDevice> CreateVerityDevice(const std::string& name,
     // Delete dangling dm-device. This can happen if apexd fails to delete it
     // while unmounting an apex.
     LOG(WARNING) << "Deleting existing dm device " << name;
-    auto result = DeleteVerityDevice(name);
+    auto result = DeleteVerityDevice(name, /* deferred= */ false);
     if (!result.ok()) {
       return result.error();
     }
@@ -578,23 +587,24 @@ Result<MountedApexData> VerifyAndTempMountPackage(
 
 }  // namespace
 
-Result<void> Unmount(const MountedApexData& data) {
+Result<void> Unmount(const MountedApexData& data, bool deferred) {
   LOG(DEBUG) << "Unmounting " << data.full_path << " from mount point "
-             << data.mount_point;
+             << data.mount_point << " deferred = " << deferred;
   // Lazily try to umount whatever is mounted.
   if (umount2(data.mount_point.c_str(), UMOUNT_NOFOLLOW) != 0 &&
       errno != EINVAL && errno != ENOENT) {
     return ErrnoError() << "Failed to unmount directory " << data.mount_point;
   }
-  // Attempt to delete the folder. If the folder is retained, other
-  // data may be incorrect.
-  if (rmdir(data.mount_point.c_str()) != 0) {
-    PLOG(ERROR) << "Failed to rmdir directory " << data.mount_point;
+
+  if (!deferred) {
+    if (rmdir(data.mount_point.c_str()) != 0) {
+      PLOG(ERROR) << "Failed to rmdir " << data.mount_point;
+    }
   }
 
   // Try to free up the device-mapper device.
   if (!data.device_name.empty()) {
-    const auto& result = DeleteVerityDevice(data.device_name);
+    const auto& result = DeleteVerityDevice(data.device_name, deferred);
     if (!result.ok()) {
       return result;
     }
@@ -604,10 +614,16 @@ Result<void> Unmount(const MountedApexData& data) {
   auto log_fn = [](const std::string& path, const std::string& /*id*/) {
     LOG(VERBOSE) << "Freeing loop device " << path << " for unmount.";
   };
-  if (!data.loop_name.empty()) {
+
+  // Since we now use LO_FLAGS_AUTOCLEAR when configuring loop devices, in
+  // theory we don't need to manually call DestroyLoopDevice here even if
+  // |deferred| is false. However we prefer to call it to ensure the invariant
+  // of SubmitStagedSession (after it's done, loop devices created for temp
+  // mount are freed).
+  if (!data.loop_name.empty() && !deferred) {
     loop::DestroyLoopDevice(data.loop_name, log_fn);
   }
-  if (!data.hashtree_loop_name.empty()) {
+  if (!data.hashtree_loop_name.empty() && !deferred) {
     loop::DestroyLoopDevice(data.hashtree_loop_name, log_fn);
   }
 
@@ -636,7 +652,7 @@ Result<void> RunVerifyFnInsideTempMount(const ApexFile& apex,
   auto cleaner = [&]() {
     if (unmount_during_cleanup) {
       LOG(DEBUG) << "Unmounting " << temp_mount_point;
-      Result<void> result = Unmount(*mount_status);
+      Result<void> result = Unmount(*mount_status, /* deferred= */ false);
       if (!result.ok()) {
         LOG(WARNING) << "Failed to unmount " << temp_mount_point << " : "
                      << result.error();
@@ -771,7 +787,7 @@ Result<void> VerifyPackageBoot(const ApexFile& apex_file) {
 // This function contains checks that might be expensive to perform, e.g. temp
 // mounting a package and reading entire dm-verity device, and shouldn't be run
 // during boot.
-Result<void> VerifyPackageInstall(const ApexFile& apex_file) {
+Result<void> VerifyPackageStagedInstall(const ApexFile& apex_file) {
   const auto& verify_package_boot_status = VerifyPackageBoot(apex_file);
   if (!verify_package_boot_status.ok()) {
     return verify_package_boot_status;
@@ -819,7 +835,7 @@ Result<ApexFile> VerifySessionDir(const int session_id) {
         "More than one APEX package found in the same session directory.");
   }
 
-  auto verified = VerifyPackages(*scan, VerifyPackageInstall);
+  auto verified = VerifyPackages(*scan, VerifyPackageStagedInstall);
   if (!verified.ok()) {
     return verified.error();
   }
@@ -939,7 +955,8 @@ Result<void> RestoreActivePackages() {
   return {};
 }
 
-Result<void> UnmountPackage(const ApexFile& apex, bool allow_latest) {
+Result<void> UnmountPackage(const ApexFile& apex, bool allow_latest,
+                            bool deferred) {
   LOG(INFO) << "Unmounting " << GetPackageId(apex.GetManifest());
 
   const ApexManifest& manifest = apex.GetManifest();
@@ -966,19 +983,15 @@ Result<void> UnmountPackage(const ApexFile& apex, bool allow_latest) {
       return Error() << "Package " << apex.GetPath() << " is active";
     }
     std::string mount_point = apexd_private::GetActiveMountPoint(manifest);
-    LOG(INFO) << "Unmounting and deleting " << mount_point;
+    LOG(INFO) << "Unmounting " << mount_point;
     if (umount2(mount_point.c_str(), UMOUNT_NOFOLLOW) != 0) {
       return ErrnoError() << "Failed to unmount " << mount_point;
-    }
-    if (rmdir(mount_point.c_str()) != 0) {
-      PLOG(ERROR) << "Could not rmdir " << mount_point;
-      // Continue here.
     }
   }
 
   // Clean up gMountedApexes now, even though we're not fully done.
   gMountedApexes.RemoveMountedApex(manifest.name(), apex.GetPath());
-  return Unmount(*data);
+  return Unmount(*data, deferred);
 }
 
 }  // namespace
@@ -1013,7 +1026,7 @@ Result<void> UnmountTempMount(const ApexFile& apex) {
       finished_unmounting = true;
     } else {
       gMountedApexes.RemoveMountedApex(manifest.name(), data->full_path, true);
-      Unmount(*data);
+      Unmount(*data, /* deferred= */ false);
     }
   }
   return {};
@@ -1303,7 +1316,8 @@ Result<void> DeactivatePackage(const std::string& full_path) {
     return apex_file.error();
   }
 
-  return UnmountPackage(*apex_file, /* allow_latest= */ true);
+  return UnmountPackage(*apex_file, /* allow_latest= */ true,
+                        /* deferred= */ false);
 }
 
 std::vector<ApexFile> GetActivePackages() {
@@ -2926,7 +2940,7 @@ int UnmountAll() {
         ret = 1;
       }
     }
-    if (auto status = Unmount(data); !status.ok()) {
+    if (auto status = Unmount(data, /* deferred= */ false); !status.ok()) {
       LOG(ERROR) << "Failed to unmount " << data.mount_point << " : "
                  << status.error();
       ret = 1;
@@ -3277,16 +3291,160 @@ android::apex::MountedApexDatabase& GetApexDatabaseForTesting() {
   return gMountedApexes;
 }
 
+// A version of apex verification that happens during non-staged APEX
+// installation.
+Result<void> VerifyPackageNonStagedInstall(const ApexFile& apex_file) {
+  const auto& verify_package_boot_status = VerifyPackageBoot(apex_file);
+  if (!verify_package_boot_status.ok()) {
+    return verify_package_boot_status;
+  }
+
+  // TODO(b/187864524): do additional checks (e.g. apk-in-apex, linkerconfig,
+  // etc.).
+  constexpr const auto kSuccessFn = [](const std::string& /*mount_point*/) {
+    return Result<void>{};
+  };
+  return RunVerifyFnInsideTempMount(apex_file, kSuccessFn, true);
+}
+
+Result<void> CheckSupportsNonStagedInstall(const ApexFile& apex) {
+  if (!apex.GetManifest().supportsrebootlessupdate()) {
+    return Error() << apex.GetPath() << " does not support non-staged update";
+  }
+  auto expected_public_key =
+      ApexFileRepository::GetInstance().GetPublicKey(apex.GetManifest().name());
+  if (!expected_public_key.ok()) {
+    return expected_public_key.error();
+  }
+  auto verity_data = apex.VerifyApexVerity(*expected_public_key);
+  if (!verity_data.ok()) {
+    return verity_data.error();
+  }
+  // Supporting non-staged install of APEXes without a hashtree is additional
+  // hassle, it's easier not to support it.
+  if (verity_data->desc->tree_size == 0) {
+    return Error() << apex.GetPath() << " does not have an embedded hash tree";
+  }
+  return {};
+}
+
+Result<std::string> ComputePackageId(const ApexFile& apex) {
+  static constexpr size_t kMaxVerityDevicesPerApexName = 3u;
+  DeviceMapper& dm = DeviceMapper::Instance();
+  std::vector<DeviceMapper::DmBlockDevice> dm_devices;
+  if (!dm.GetAvailableDevices(&dm_devices)) {
+    return Error() << "Failed to list dm devices";
+  }
+  size_t devices = 0;
+  size_t next_minor = 1;
+  for (const auto& dm_device : dm_devices) {
+    std::string_view dm_name(dm_device.name());
+    // Format is <module_name>@<version_code>[_<minor>]
+    if (!ConsumePrefix(&dm_name, apex.GetManifest().name())) {
+      continue;
+    }
+    devices++;
+    auto pos = dm_name.find_last_of('_');
+    if (pos == std::string_view::npos) {
+      continue;
+    }
+    size_t minor;
+    if (!ParseUint(std::string(dm_name.substr(pos + 1)), &minor)) {
+      return Error() << "Unexpected dm device name " << dm_device.name();
+    }
+    if (next_minor < minor + 1) {
+      next_minor = minor + 1;
+    }
+  }
+  if (devices > kMaxVerityDevicesPerApexName) {
+    return Error() << "There are too many (" << devices
+                   << ") dm block devices associated with package "
+                   << apex.GetManifest().name();
+  }
+  std::string id =
+      GetPackageId(apex.GetManifest()) + "_" + std::to_string(next_minor);
+  return std::move(id);
+}
+
 Result<ApexFile> InstallPackage(const std::string& package_path) {
   LOG(INFO) << "Installing " << package_path;
-  auto apex = ApexFile::Open(package_path);
-  if (!apex.ok()) {
-    return apex;
+  auto temp_apex = ApexFile::Open(package_path);
+  if (!temp_apex.ok()) {
+    return temp_apex.error();
   }
-  if (!apex->GetManifest().supportsrebootlessupdate()) {
-    return Error() << package_path << " does not support non-staged update";
+
+  const std::string& module_name = temp_apex->GetManifest().name();
+
+  // Do a quick check if this APEX can be installed without a reboot.
+  // Note that passing this check doesn't guarantee that APEX will be
+  // successfully installed.
+  if (auto res = CheckSupportsNonStagedInstall(*temp_apex); !res.ok()) {
+    return res.error();
   }
-  return Error() << "TODO(b/187864524): implement";
+
+  // Don't allow non-staged update if there are no active versions of this APEX.
+  auto cur_mounted_data = gMountedApexes.GetLatestMountedApex(module_name);
+  if (!cur_mounted_data.has_value()) {
+    return Error() << "No active version found for package " << module_name;
+  }
+
+  auto cur_apex = ApexFile::Open(cur_mounted_data->full_path);
+  if (!cur_apex.ok()) {
+    return cur_apex.error();
+  }
+
+  // 1. Verify that APEX is correct. This is a heavy check that involves
+  // mounting an APEX on a temporary mount point and reading the entire
+  // dm-verity block device.
+  if (auto verify = VerifyPackageNonStagedInstall(*temp_apex); !verify.ok()) {
+    return verify.error();
+  }
+
+  // 2. Compute params for mounting new apex.
+  auto new_id = ComputePackageId(*temp_apex);
+  if (!new_id.ok()) {
+    return new_id.error();
+  }
+
+  // 2. Unmount currently active APEX.
+  if (auto res = UnmountPackage(*cur_apex, /* allow_latest= */ true,
+                                /* deferred= */ true);
+      !res.ok()) {
+    return res.error();
+  }
+
+  // TODO(b/188713178): handle errors and cleanup at different stages.
+
+  // 3. Hard link to final destination.
+  std::string target_file = StringPrintf(
+      "%s/%s.apex", gConfig->active_apex_data_dir, (*new_id).c_str());
+  // At this point it should be safe to hard link |temp_apex| to
+  // |params->target_file|. In case reboot happens during one of the stages
+  // below, then on next boot apexd will pick up the new verified APEX.
+  if (link(package_path.c_str(), target_file.c_str()) != 0) {
+    return ErrnoError() << "Failed to link " << package_path << " to "
+                        << target_file;
+  }
+
+  auto new_apex = ApexFile::Open(target_file);
+  if (!new_apex.ok()) {
+    return new_apex.error();
+  }
+
+  // 4. And activate new one.
+  if (auto res = ActivatePackageImpl(*new_apex, *new_id); !res.ok()) {
+    return res.error();
+  }
+
+  // 4. Now we can ulink old APEX if it's not pre-installed.
+  if (!ApexFileRepository::GetInstance().IsPreInstalledApex(*cur_apex)) {
+    if (unlink(cur_mounted_data->full_path.c_str()) != 0) {
+      PLOG(ERROR) << "Failed to unlink " << cur_mounted_data->full_path;
+    }
+  }
+
+  // TODO(b/188713178): update apex-info-list.xml
+  return new_apex;
 }
 
 }  // namespace apex
